pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gbif-downloader"
version = "2.0.0"
description = "Download and filter GBIF biodiversity occurrence data with CLI and GUI interfaces"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Francesco Mensa"}
]
keywords = [
    "gbif",
    "biodiversity",
    "occurrence",
    "museum",
    "specimen",
    "taxonomy",
    "entomology",
    "ecology"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "openpyxl>=3.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "geojson>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "responses>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gbif-download = "gbif_downloader.cli:main"
gbif-gui = "gbif_downloader.gui:main"

[project.urls]
Homepage = "https://github.com/yourusername/gbif-downloader"
Documentation = "https://github.com/yourusername/gbif-downloader#readme"
Repository = "https://github.com/yourusername/gbif-downloader"
Issues = "https://github.com/yourusername/gbif-downloader/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["gbif_downloader*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=gbif_downloader --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]
